+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 2.1.0
+- Remove H265 codec and deprecated frame-marking RTP extension
+- Remove H264-SVC codec
+- `Router`: Add `UpdateMediaCodecs()` method to dynamically change Router's RTP capabilities
+- add version support for mediasoup C++ subprocess
+
 ### 2.0.3
 
 - feat: Add initial AV1 codec support